example and doc'
